requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-cloud-firestore==2.20.2
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.3